tests/wpt/web-platform-tests/app-history/navigate/return-value/navigate-interrupted-within-onnavigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  let result2;
  appHistory.onnavigate = t.step_func(e => {
    if (e.info == 1) {
      result2 = appHistory.navigate("#2", { info: 2 });
      assert_true(e.defaultPrevented);
    }
  });

  const result1 = appHistory.navigate("#1", { info: 1 });

  assert_equals(appHistory.entries().length, 2);
  assert_array_equals(appHistory.entries().map(e => (new URL(e.url)).hash), ["", "#2"]);

  await assertBothRejectDOM(t, result1, "AbortError");
  await assertBothFulfill(t, result2, appHistory.current);
}, "if navigate() is called inside onnavigate, the previous navigation and navigate event are cancelled");
</script>
